refactor: package relocations together
