+++ HospitalSystem.ConsoleUI/Director/EquipmentUI.cs
@@ -2,6 +2,7 @@ using System.Text.RegularExpressions;
 using HospitalSystem.Core;
 using HospitalSystem.Core.Rooms;
 using HospitalSystem.Core.Equipment;
+using HospitalSystem.Core.Equipment.Relocations;
 
 namespace HospitalSystem.ConsoleUI.Director;
